added ability to generate all spreadsheets
