+++ main.py
@@ -1,71 +1,8 @@
 import gspread
-from openpyxl import load_workbook
 
 import config.user_info as usr
 from config.gsheets import client
-
-# (1) Use openpyxl to create templates that can be filled with data
-
-class Template:
-    """
-    Represents an Excel form template, that can additionally
-    be filled with data.
-    """
-
-    def _save_file(self, wb, filepath):
-        """
-        Saves file if filepath is specified.
-        """
-        if filepath:
-            wb.save(filepath)
-
-    def get_empty(self, filepath=None):
-        """
-        Returns empty template.
-        Optionally saves as an Excel workbook if filepath is
-        specified.
-        """
-        wb = load_workbook(filepath='template.xlsx')
-        self._save_file(wb, filepath)
-        return wb
-
-    def get_new(self, filepath=None):
-        """
-        Returns new template with org_name, bookkeeper,
-        treasurer and address stub variables filled.
-        Optionally saves as an Excel workbook if filepath is
-        specified.
-        """
-        wb = load_workbook(filepath = 'template.xlsx')
-        ws = wb.active
-
-        ws[usr.tmp_vars['org_name']] = usr.org_name
-        ws[usr.tmp_vars['bookkeeper']] = usr.bookkeeper
-        ws[usr.tmp_vars['treasurer']] = usr.treasurer
-        ws[usr.tmp_vars['address']] = usr.address
-
-        self._save_file(wb, filepath)
-        return wb
-
-    def get_completed(self, data_dict, filepath=None):
-        """
-        Return completed form given a dictionary of values
-        to plug into the template.
-        Optionally saves as an Excel workbook if filepath is
-        specified.
-
-        data_dict: Dictionary of user specific data for each form, where
-        keys are strings corresponding to usr.tmp_vars key names. 
-        """
-        wb = self.get_new()
-        ws = wb.active
-        for varname in data_dict:
-            # only available data will be filled in
-            ws[usr.tmp_vars[varname]] = data_dict[varname]
-        self._save_file(wb, filepath)
-        return wb
-
-# (2) get data for each user
+from template import FormTemplate
 
 def get_user_dict():
     """
@@ -95,11 +32,11 @@ def get_new_row_range(ws):
     """
     row_range = []
     for row in range(2, ws.row_count):
-        # if timestamp is missing data assumed to be missing
-        if not ws.cell(row, usr.rmbs_cols['date']):
+        # if timestamp is missing, data assumed to be missing
+        if not ws.cell(row, usr.rmbs_cols['date']).value:
             break
         # has form is None or false if no form has been created yet
-        if not ws.cell(row, usr.rmbs_cols['has_form']):
+        if not ws.cell(row, usr.rmbs_cols['has_form']).value:
             row_range.append(row)
     return row_range
 
@@ -117,10 +54,7 @@ def get_rmbs_dict(row_data):
     converts a list of reimbursements data for a given row to a dictionary
     of that data.
     """
-    new_data = {}
-    for var in usr.rmbs_cols:
-        new_data[var] = row_data[rmbs_cols[var] - 1]
-    return new_data
+    return {var: row_data[usr.rmbs_cols[var] - 1] for var in usr.rmbs_cols}
 
 def get_data_dict(ws, row_number, user_data):
     """
@@ -128,15 +62,15 @@ def get_data_dict(ws, row_number, user_data):
     the row number to access and a dictionary of data for each user,
     returns a dictionary of all data associated with that row of reimbursements.
     """
-    rmbs_data = get_rmbs_dict(ws.row_values(row))
-    username = rmbs_data[usr.rmbs_cols['username']]
+    rmbs_data = get_rmbs_dict(ws.row_values(row_number))
+    username = rmbs_data['username']
     try:
         user_values = dict(user_data[username], **rmbs_data)
     except KeyError:
         raise ValueError("User {} has not submitted their info to the contacts form".format(username))
-    return {var: user_values.get(var, None) for var in usr.tmp_vars}
+    return {var: user_values[var] for var in usr.tmp_vars if var in user_values}
 
-def get_worksheet_data(ws, row_range):
+def get_rmbs_data(ws, row_range):
     """
     given a gspread worksheet representing the form accepting new 
     reimbursements and list of row range, returns a list of dictionaries 
@@ -145,9 +79,23 @@ def get_worksheet_data(ws, row_range):
     user_data = get_user_dict()
     return [get_data_dict(ws, row, user_data) for row in row_range]
 
-# (3) make spreadsheets for users and update converted rows
-
 def make_spreadsheets():
     """
     """
     rmbs = client.open_by_url(usr.rmbs_url).sheet1
+    row_range = get_new_row_range(rmbs)
+    print "loading data for rows", row_range
+    all_data = get_rmbs_data(rmbs, row_range)
+    for data_dict in all_data:
+        # date structured as: "%m/%d/%Y %H:%M:%S"
+        date = data_dict['date'].split()[0].replace('/', '')
+        # name structured as "Angelina Li"
+        name = data_dict['stu_name'].split()[0].lower()
+        filepath = "output/user/{name}_{date}.xlsx".format(
+                    name=name,
+                    date=date)
+        print "generating form for {}".format(name)
+        FormTemplate().get_completed(data_dict, filepath)
+
+make_spreadsheets()
+
